agent/requirements.txt
python-socketio[client]
requests
python-dotenv
pyngrok
jupyterlab
